Add files via upload
A pattern or sequence detector is a sequential circuit (/state machine) which takes an input bit stream and generates an output '1' when the desired pattern is detected. The basic finite state machine can be of Melay type or Moore type. In Moore machine output values are determined only by its current state while in  a Mealy machine, output depends on its current state and the current inputs. Sequence detector is of two types:
    Overlapping
    Non-Overlapping
In an overlapping sequence detector, the last bit of one sequence becomes the first bit of the next sequence. However, in a non-overlapping sequence detector, the last bit of one sequence does not become the first bit of the next sequence.